Added constant buffers component.
