+++ Code/Engine/RenderSystem/RenderSystemDX10/RenderSystemDX10.cpp
@@ -147,110 +147,10 @@ void LcRenderSystemDX10::Create(void* windowHandle, LcWinMode winMode, bool inVS
 
 	d3dDevice->RSSetViewports(1, &viewPort);
 
-	// define constant buffers
-	struct VS_MATRIX_BUFFER
-	{
-		LcMatrix4 mat;
-	};
-	VS_MATRIX_BUFFER matData;
-
-	struct VS_COLORS_BUFFER
-	{
-		LcColor4 colors[4];
-	};
-	VS_COLORS_BUFFER colorsData;
-
-	struct VS_CUSTOM_UV_BUFFER
-	{
-		LcVector4 uv[4];
-	};
-	VS_CUSTOM_UV_BUFFER uvData;
-
-	struct VS_FRAME_ANIM2D_BUFFER
-	{
-		LcVector4 animData;
-	};
-	VS_FRAME_ANIM2D_BUFFER anim2dData;
-
-	struct VS_SETTINGS_BUFFER
-	{
-		LcVector4 worldTint;
-	};
-	VS_SETTINGS_BUFFER settingsData;
-
-	D3D10_BUFFER_DESC cbDesc{};
-	cbDesc.ByteWidth = sizeof(VS_MATRIX_BUFFER);
-	cbDesc.Usage = D3D10_USAGE_DEFAULT;
-	cbDesc.BindFlags = D3D10_BIND_CONSTANT_BUFFER;
-
-	D3D10_SUBRESOURCE_DATA subResData{};
-	subResData.pSysMem = &matData;
-
-	// create constant buffers
-	matData.mat = OrthoMatrix(LcSize(width, height), 1.0f, -1.0f);
-	if (FAILED(d3dDevice->CreateBuffer(&cbDesc, &subResData, projMatrixBuffer.GetAddressOf())))
-	{
-		throw std::exception("LcRenderSystemDX10(): Cannot create constant buffer");
-	}
-
-	matData.mat = IdentityMatrix();
-	if (FAILED(d3dDevice->CreateBuffer(&cbDesc, &subResData, transMatrixBuffer.GetAddressOf())))
-	{
-		throw std::exception("LcRenderSystemDX10(): Cannot create constant buffer");
-	}
-
+	// init constant buffers
 	cameraPos = LcVector3(width / 2.0f, height / 2.0f, 0.0f);
 	cameraTarget = LcVector3(cameraPos.x, cameraPos.y, 1.0f);
-	matData.mat = LookAtMatrix(cameraPos, cameraTarget);
-	if (FAILED(d3dDevice->CreateBuffer(&cbDesc, &subResData, viewMatrixBuffer.GetAddressOf())))
-	{
-		throw std::exception("LcRenderSystemDX10(): Cannot create constant buffer");
-	}
-
-	subResData.pSysMem = &colorsData;
-	colorsData.colors[0] = LcDefaults::White4;
-	colorsData.colors[1] = LcDefaults::White4;
-	colorsData.colors[2] = LcDefaults::White4;
-	colorsData.colors[3] = LcDefaults::White4;
-	if (FAILED(d3dDevice->CreateBuffer(&cbDesc, &subResData, colorsBuffer.GetAddressOf())))
-	{
-		throw std::exception("LcRenderSystemDX10(): Cannot create constant buffer");
-	}
-
-	subResData.pSysMem = &uvData;
-	uvData.uv[1] = To4(LcVector2(1.0, 0.0));
-	uvData.uv[2] = To4(LcVector2(1.0, 1.0));
-	uvData.uv[0] = To4(LcVector2(0.0, 0.0));
-	uvData.uv[3] = To4(LcVector2(0.0, 1.0));
-	if (FAILED(d3dDevice->CreateBuffer(&cbDesc, &subResData, customUvBuffer.GetAddressOf())))
-	{
-		throw std::exception("LcRenderSystemDX10(): Cannot create constant buffer");
-	}
-
-	cbDesc.ByteWidth = sizeof(VS_FRAME_ANIM2D_BUFFER);
-	subResData.pSysMem = &anim2dData;
-	anim2dData.animData = LcVector4(1.0, 1.0, 0.0, 0.0);
-	if (FAILED(d3dDevice->CreateBuffer(&cbDesc, &subResData, frameAnimBuffer.GetAddressOf())))
-	{
-		throw std::exception("LcRenderSystemDX10(): Cannot create constant buffer");
-	}
-
-	cbDesc.ByteWidth = sizeof(VS_SETTINGS_BUFFER);
-	subResData.pSysMem = &settingsData;
-	settingsData.worldTint = LcVector4(1.0, 1.0, 1.0, 1.0);
-	if (FAILED(d3dDevice->CreateBuffer(&cbDesc, &subResData, settingsBuffer.GetAddressOf())))
-	{
-		throw std::exception("LcRenderSystemDX10(): Cannot create constant buffer");
-	}
-
-	// set buffers
-	d3dDevice->VSSetConstantBuffers(0, 1, projMatrixBuffer.GetAddressOf());
-	d3dDevice->VSSetConstantBuffers(1, 1, viewMatrixBuffer.GetAddressOf());
-	d3dDevice->VSSetConstantBuffers(2, 1, transMatrixBuffer.GetAddressOf());
-	d3dDevice->VSSetConstantBuffers(3, 1, colorsBuffer.GetAddressOf());
-	d3dDevice->VSSetConstantBuffers(4, 1, customUvBuffer.GetAddressOf());
-	d3dDevice->VSSetConstantBuffers(5, 1, frameAnimBuffer.GetAddressOf());
-	d3dDevice->VSSetConstantBuffers(6, 1, settingsBuffer.GetAddressOf());
+	constBuffers.Init(d3dDevice.Get(), cameraPos, cameraTarget, width, height);
 
 	// create blend state
 	D3D10_BLEND_DESC blendStateDesc;
@@ -333,17 +233,11 @@ void LcRenderSystemDX10::Shutdown()
 	visual2DRenders.clear();
 	widgetRender = nullptr;
 
+	constBuffers.Destroy();
 	depthStencilView.Reset();
 	depthStencil.Reset();
 	rasterizerState.Reset();
 	blendState.Reset();
-	settingsBuffer.Reset();
-	frameAnimBuffer.Reset();
-	customUvBuffer.Reset();
-	colorsBuffer.Reset();
-	transMatrixBuffer.Reset();
-	projMatrixBuffer.Reset();
-	viewMatrixBuffer.Reset();
 	renderTargetView.Reset();
 	swapChain.Reset();
 	d3dDevice.Reset();
@@ -370,10 +264,11 @@ void LcRenderSystemDX10::Subscribe(const LcAppContext& context)
 
 	context.world.onTintChanged.AddListener([this](LcColor3 globalTint)
 	{
-		if (settingsBuffer)
+		if (constBuffers.settingsBuffer)
 		{
-			LcVector4 settings = To4(globalTint);
-			d3dDevice->UpdateSubresource(settingsBuffer.Get(), 0, NULL, &settings, 0, 0);
+			VS_SETTINGS_BUFFER settings;
+			settings.worldTint = To4(globalTint);
+			d3dDevice->UpdateSubresource(constBuffers.settingsBuffer.Get(), 0, NULL, &settings, 0, 0);
 		}
 	});
 }
@@ -386,7 +281,7 @@ void LcRenderSystemDX10::Update(float deltaSeconds, const LcAppContext& context)
 void LcRenderSystemDX10::UpdateCamera(float deltaSeconds, LcVector3 newPos, LcVector3 newTarget)
 {
 	auto viewMatrix = LookAtMatrix(newPos, newTarget);
-	d3dDevice->UpdateSubresource(viewMatrixBuffer.Get(), 0, NULL, &viewMatrix, 0, 0);
+	d3dDevice->UpdateSubresource(constBuffers.viewMatrixBuffer.Get(), 0, NULL, &viewMatrix, 0, 0);
 }
 
 void LcRenderSystemDX10::Render(const LcAppContext& context)
@@ -483,7 +378,7 @@ void LcRenderSystemDX10::Resize(int width, int height, const LcAppContext& conte
 
 		// update projection matrix
 		LcMatrix4 proj = OrthoMatrix(LcSize(width, height), 1.0f, -1.0f);
-		d3dDevice->UpdateSubresource(projMatrixBuffer.Get(), 0, NULL, &proj, 0, 0);
+		d3dDevice->UpdateSubresource(constBuffers.projMatrixBuffer.Get(), 0, NULL, &proj, 0, 0);
 
 		renderSystemSize = newViewportSize;
 	}
