+++ Projects/Windows/RenderSystemDX10/RenderSystemDX10.vcxproj.filters
@@ -57,6 +57,9 @@
     <ClInclude Include="..\..\..\Code\Engine\RenderSystem\RenderSystemDX10\VisualsDX10.h">
       <Filter>Header Files\RenderSystemDX10</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\..\Code\Engine\RenderSystem\RenderSystemDX10\ConstantBuffersDX10.h">
+      <Filter>Header Files\RenderSystemDX10</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="dllmain.cpp">
@@ -87,6 +90,9 @@
       <Filter>Source Files\RenderSystemDX10</Filter>
     </ClCompile>
     <ClCompile Include="..\..\..\Code\Engine\RenderSystem\RenderSystemDX10\VisualsDX10.cpp">
+      <Filter>Source Files\RenderSystemDX10</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\Code\Engine\RenderSystem\RenderSystemDX10\ConstantBuffersDX10.cpp">
       <Filter>Source Files\RenderSystemDX10</Filter>
     </ClCompile>
   </ItemGroup>
